Simplify fixture structure
