+++ runtimes/alibi-explain/tests/conftest.py
@@ -30,36 +30,9 @@ from mlserver_alibi_explain.runtime import AlibiExplainRuntime, AlibiExplainRunt
 
 from .helpers.tf_model import get_tf_mnist_model_uri, TFMNISTModel
 from .helpers.run_async import run_async_as_sync
-from prometheus_client.registry import REGISTRY
-from starlette_exporter import PrometheusMiddleware
 
 TESTS_PATH = Path(os.path.dirname(__file__))
 _ANCHOR_IMAGE_DIR = TESTS_PATH / ".data" / "mnist_anchor_image"
-
-
-@pytest.fixture(autouse=True)
-def prometheus_registry() -> Iterable:
-    """
-    Fixture used to ensure the registry is cleaned on each run.
-    Otherwise, `py-grpc-prometheus` will complain that metrics already exist.
-
-    TODO: Open issue in `py-grpc-prometheus` to check whether a metric exists
-    before creating it.
-    For an example on how to do this, see `starlette_exporter`'s implementation
-
-        https://github.com/stephenhillier/starlette_exporter/blob/947d4d631dd9a6a8c1071b45573c5562acba4834/starlette_exporter/middleware.py#L67
-    """
-    # NOTE: Since the `REGISTRY` object is global, this fixture is NOT
-    # thread-safe!!
-    collectors = list(REGISTRY._collector_to_names.keys())
-    for collector in collectors:
-        REGISTRY.unregister(collector)
-
-    # Clean metrics from `starlette_exporter` as well, as otherwise they won't
-    # get re-created
-    PrometheusMiddleware._metrics.clear()
-
-    yield REGISTRY
 
 
 @pytest.fixture
