+++ tests/conftest.py
@@ -3,6 +3,7 @@ import os
 import shutil
 import asyncio
 import logging
+import glob
 
 from starlette_exporter import PrometheusMiddleware
 from prometheus_client.registry import REGISTRY, CollectorRegistry
@@ -22,7 +23,7 @@ from mlserver.env import Environment
 from mlserver import types, Settings, ModelSettings
 
 from .fixtures import SumModel, ErrorModel, SimpleModel
-from .utils import _pack, _get_tarball_name
+from .utils import RESTClient, get_available_ports, _pack, _get_tarball_name
 
 TESTS_PATH = os.path.dirname(__file__)
 TESTDATA_PATH = os.path.join(TESTS_PATH, "testdata")
@@ -76,7 +77,6 @@ def logger():
     return logger
 
 
-@pytest.fixture(autouse=True)
 def prometheus_registry() -> CollectorRegistry:
     """
     Fixture used to ensure the registry is cleaned on each run.
@@ -88,6 +88,8 @@ def prometheus_registry() -> CollectorRegistry:
 
         https://github.com/stephenhillier/starlette_exporter/blob/947d4d631dd9a6a8c1071b45573c5562acba4834/starlette_exporter/middleware.py#L67
     """
+    yield REGISTRY
+
     # NOTE: Since the `REGISTRY` object is global, this fixture is NOT
     # thread-safe!!
     collectors = list(REGISTRY._collector_to_names.keys())
@@ -97,8 +99,6 @@ def prometheus_registry() -> CollectorRegistry:
     # Clean metrics from `starlette_exporter` as well, as otherwise they won't
     # get re-created
     PrometheusMiddleware._metrics.clear()
-
-    yield REGISTRY
 
 
 @pytest.fixture
@@ -250,3 +250,53 @@ async def inference_pool(settings: Settings) -> InferencePool:
     yield pool
 
     await pool.close()
+
+
+@pytest.fixture
+def _mlserver_settings(settings: Settings):
+    """
+    This is an indirect fixture used to tweak the standard settings ONLY when
+    the `mlserver` fixture is used.
+    You shouldn't need to use this fixture directly.
+    """
+    http_port, grpc_port, metrics_port = get_available_ports(3)
+    settings.http_port = http_port
+    settings.grpc_port = grpc_port
+    settings.metrics_port = metrics_port
+    settings.metrics_dir = str(tmp_path)
+
+    return settings
+
+
+@pytest.fixture
+async def mlserver(
+    _mlserver_settings: Settings,
+    sum_model_settings: ModelSettings,
+    prometheus_registry: CollectorRegistry,
+):
+    server = MLServer(_mlserver_settings)
+
+    # Start server without blocking, and cancel afterwards
+    server_task = asyncio.create_task(server.start())
+
+    # Load sample model
+    await server._model_registry.load(sum_model_settings)
+
+    yield server
+
+    await server.stop()
+    await server_task
+
+    pattern = os.path.join(_mlserver_settings.metrics_dir, "*.db")
+    prom_files = glob.glob(pattern)
+    assert not prom_files
+
+
+@pytest.fixture
+async def rest_client(mlserver: MLServer, settings: Settings):
+    http_server = f"{settings.host}:{settings.http_port}"
+    client = RESTClient(http_server)
+
+    yield client
+
+    await client.close()
